Minor code cleanup

+++ src/main/java/com/adbazaar/dto/book/BookShortDetails.java
@@ -28,7 +28,7 @@ public class BookShortDetails {
 
     private String format;
 
-    private BigDecimal rate;
+    private Double rate;
 
     private Integer quantity;
